Abstract away core functionality of storage backed instance
