+++ src/main/java/org/commcare/cases/instance/CaseInstanceTreeElement.java
@@ -1,16 +1,10 @@
 package org.commcare.cases.instance;
 
 import org.commcare.cases.model.Case;
-import org.commcare.cases.util.StorageBackedTreeRoot;
-import org.javarosa.core.model.data.IAnswerData;
 import org.javarosa.core.model.instance.AbstractTreeElement;
-import org.javarosa.core.model.instance.TreeElement;
-import org.javarosa.core.model.instance.TreeReference;
-import org.javarosa.core.model.instance.utils.ITreeVisitor;
 import org.javarosa.core.services.storage.IStorageIterator;
 import org.javarosa.core.services.storage.IStorageUtilityIndexed;
 import org.javarosa.core.util.DataUtil;
-import org.javarosa.core.util.Interner;
 import org.javarosa.model.xform.XPathReference;
 import org.javarosa.xpath.expr.XPathPathExpr;
 
@@ -24,7 +18,7 @@ import java.util.Vector;
  *
  * @author ctsims
  */
-public class CaseInstanceTreeElement extends StorageBackedTreeRoot<CaseChildElement> {
+public class CaseInstanceTreeElement extends StorageInstanceTreeElement<CaseChildElement> {
 
     public static final String MODEL_NAME = "casedb";
 
@@ -35,212 +29,30 @@ public class CaseInstanceTreeElement extends StorageBackedTreeRoot<CaseChildElem
     private final static XPathPathExpr CASE_STATUS_EXPR = XPathReference.getPathExpr("@status");
     private final static XPathPathExpr CASE_INDEX_EXPR = XPathReference.getPathExpr("index/*");
 
-    private AbstractTreeElement instanceRoot;
-
-    protected final IStorageUtilityIndexed storage;
-    protected Vector<CaseChildElement> cases;
-    protected final Interner<TreeElement> treeCache = new Interner<>();
-    private Interner<String> stringCache;
-
-    private int numRecords = -1;
-    private TreeReference cachedRef = null;
-
     public CaseInstanceTreeElement(AbstractTreeElement instanceRoot,
                                    IStorageUtilityIndexed storage) {
-        this.instanceRoot = instanceRoot;
-        this.storage = storage;
-    }
-
-    /**
-     * Rebase assigns this tree element to a new root instance node.
-     *
-     * Used to migrate the already created tree structure to a new instance connector.
-     *
-     * @param instanceRoot The root of the new tree that this element should be a part of
-     */
-    public void rebase(AbstractTreeElement instanceRoot) {
-        this.instanceRoot = instanceRoot;
-        expireCachedRef();
+        super(instanceRoot, storage, MODEL_NAME, "case");
     }
 
     @Override
-    public boolean isLeaf() {
-        return false;
-    }
-
-    @Override
-    public boolean isChildable() {
-        return false;
-    }
-
-    @Override
-    public String getInstanceName() {
-        return instanceRoot.getInstanceName();
-    }
-
-    public void attachStringCache(Interner<String> stringCache) {
-        this.stringCache = stringCache;
-    }
-
-    @Override
-    public CaseChildElement getChild(String name, int multiplicity) {
-        if ((multiplicity == TreeReference.INDEX_TEMPLATE) &&
-                "case".equals(name)) {
-            return CaseChildElement.buildCaseChildTemplate(this);
-        }
-
-        //name is always "case", so multiplicities are the only relevant component here
-        if (name.equals("case")) {
-            getCases();
-            if (cases.isEmpty()) {
-                //If we have no cases, we still need to be able to return a template element so as to not
-                //break xpath evaluation
-                return CaseChildElement.buildCaseChildTemplate(this);
-            }
-            return cases.elementAt(multiplicity);
-        }
-        return null;
-    }
-
-    @Override
-    public Vector<CaseChildElement> getChildrenWithName(String name) {
-        if (name.equals("case")) {
-            getCases();
-            return cases;
-        } else {
-            return new Vector<>();
-        }
-    }
-
-    @Override
-    public boolean hasChildren() {
-        return getNumChildren() > 0;
-    }
-
-    @Override
-    public int getNumChildren() {
-        if (numRecords == -1) {
-            numRecords = storage.getNumRecords();
-        }
-        return numRecords;
-    }
-
-    @Override
-    public CaseChildElement getChildAt(int i) {
-        getCases();
-        return cases.elementAt(i);
-    }
-
-    protected synchronized void getCases() {
-        if (cases != null) {
+    protected synchronized void loadElements() {
+        if (elements != null) {
             return;
         }
         objectIdMapping = new Hashtable<>();
-        cases = new Vector<>();
+        elements = new Vector<>();
         int mult = 0;
         for (IStorageIterator i = storage.iterate(); i.hasMore(); ) {
             int id = i.nextID();
-            cases.addElement(new CaseChildElement(this, id, null, mult));
+            elements.addElement(new CaseChildElement(this, id, null, mult));
             objectIdMapping.put(DataUtil.integer(id), DataUtil.integer(mult));
             mult++;
         }
     }
 
     @Override
-    public boolean isRepeatable() {
-        return false;
-    }
-
-    @Override
-    public boolean isAttribute() {
-        return false;
-    }
-
-    @Override
-    public int getChildMultiplicity(String name) {
-        //All children have the same name;
-        if (name.equals("case")) {
-            return this.getNumChildren();
-        } else {
-            return 0;
-        }
-    }
-
-    @Override
-    public boolean isRelevant() {
-        return true;
-    }
-
-    @Override
-    public void accept(ITreeVisitor visitor) {
-        visitor.visit(this);
-    }
-
-    @Override
-    public int getAttributeCount() {
-        return 0;
-    }
-
-    @Override
-    public String getAttributeNamespace(int index) {
-        return null;
-    }
-
-    @Override
-    public String getAttributeName(int index) {
-        return null;
-    }
-
-    @Override
-    public String getAttributeValue(int index) {
-        return null;
-    }
-
-    @Override
-    public CaseChildElement getAttribute(String namespace, String name) {
-        return null;
-    }
-
-    @Override
-    public String getAttributeValue(String namespace, String name) {
-        return null;
-    }
-
-    @Override
-    public TreeReference getRef() {
-        if (cachedRef == null) {
-            cachedRef = TreeReference.buildRefFromTreeElement(this);
-        }
-        return cachedRef;
-    }
-
-    private void expireCachedRef() {
-        cachedRef = null;
-    }
-
-    @Override
-    public String getName() {
-        return MODEL_NAME;
-    }
-
-    @Override
-    public int getMult() {
-        return 0;
-    }
-
-    @Override
-    public AbstractTreeElement getParent() {
-        return instanceRoot;
-    }
-
-    @Override
-    public IAnswerData getValue() {
-        return null;
-    }
-
-    @Override
-    public int getDataType() {
-        return 0;
+    protected CaseChildElement getChildTemplate(StorageInstanceTreeElement parent) {
+        return CaseChildElement.buildCaseChildTemplate(this);
     }
 
     @Override
@@ -257,26 +69,8 @@ public class CaseInstanceTreeElement extends StorageBackedTreeRoot<CaseChildElem
         return filter;
     }
 
-    @Override
-    public String getNamespace() {
-        return null;
-    }
-
-    public String intern(String s) {
-        if (stringCache == null) {
-            return s;
-        } else {
-            return stringCache.intern(s);
-        }
-    }
-
     protected Case getCase(int recordId) {
         return (Case)storage.read(recordId);
-    }
-
-    @Override
-    protected String getChildHintName() {
-        return "case";
     }
 
     @Override
@@ -291,15 +85,5 @@ public class CaseInstanceTreeElement extends StorageBackedTreeRoot<CaseChildElem
         indices.put(CASE_INDEX_EXPR, Case.INDEX_CASE_INDEX_PRE);
 
         return indices;
-    }
-
-    @Override
-    protected IStorageUtilityIndexed<?> getStorage() {
-        return storage;
-    }
-
-    @Override
-    protected void initStorageCache() {
-        getCases();
     }
 }
